+++ python_search/events/ranking_generated/__init__.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 from typing import List, Union
-from python_search.config import DataConfig
+from python_search.configuration.data_config import DataConfig
 from python_search.data_collector import GenericDataCollector
 
 import uuid
