+++ python_search/events/run_performed/dataset.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 
-from python_search.config import DataConfig
+from python_search.configuration.data_config import DataConfig
 from python_search.data_collector import GenericDataCollector
 
 class EntryExecutedDataset:
